requirements.txt
tensorflow-gpu==2.2.0
pandas
tqdm
scikit-learn
prettytable
matplotlib
seaborn
tabulate
pyqt5
pycm
numpy==1.21